shop/requirements.txt
django
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pillow
bootstrap4
crispy-bootstrap4
python-dotenv
celery
reportlab
redis
requests
openpyxl
djangorestframework
sentry-sdk
drf_spectacular
pytest-django
pytest
factory_boy
model-bakery
localstack
boto3
django-storages
